requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
audioplayer==0.6
audioread==3.0.1
Automat==22.10.0
beautifulsoup4==4.12.3
black==23.10.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.6
comm==0.2.1
constantly==23.10.4
contourpy==1.1.1
cryptography==42.0.5
cssselect==1.2.0
cvlib==0.2.7
cycler==0.12.1
Cython==3.0.4
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.1.1
executing==2.0.1
ez-setup==0.9
filelock==3.13.3
fire==0.5.0
Flask==3.0.3
flatbuffers==24.3.7
fonttools==4.43.1
fsspec==2024.3.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
geojson==2.5.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.1
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.4
imageio==2.31.6
imantics==0.1.12
imbalanced-learn==0.13.0
imblearn==0.0
imgaug==0.4.0
imutils==0.5.4
incremental==22.10.0
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel==6.29.2
ipympl==0.9.3
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keras==3.1.1
kiwisolver==1.4.5
labelme2coco==0.2.4
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2
nltk==3.8.1
numba==0.59.1
numpy==1.26.1
omegaconf==2.3.0
openai==1.16.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.10.0
packaging==23.2
pandas==2.2.1
parsel==1.9.0
parso==0.8.3
pathspec==0.11.2
pefile==2023.2.7
pillow==10.3.0
pixellib==0.7.1
platformdirs==3.11.0
pooch==1.8.1
portalocker==2.8.2
progressbar==2.5
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.7
pydot==1.4.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyowm==3.3.0
pyparsing==3.1.1
pyperclip==1.8.2
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
queuelib==1.6.2
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.10.6
safetensors==0.4.2
sahi==0.11.14
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.3
Scrapy==2.11.1
seaborn==0.13.2
service-identity==24.1.0
setuptools==68.2.2
shapely==2.0.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.3
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.9.26
tldextract==5.1.2
tokenizers==0.15.2
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
tzdata==2024.1
ultralytics==8.2.4
urllib3==2.0.7
w3lib==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
widgetsnbextension==4.0.10
wifi-qrcode-generator==1.0.1
wifi_qrcode==2.1.0
wikipedia==1.4.0
wrapt==1.16.0
xgboost==2.0.3
xmljson==0.2.1
yacs==0.1.8
zope.interface==6.2
